# contracts/openapi/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Lpsaring API Contract
  version: 1.0.0
  description: |
    Source of truth kontrak API domain prioritas (Auth, Profile, Devices, Transactions).
    Ruang lingkup ini sengaja fokus untuk mengurangi drift backend-frontend pada flow kritikal.
    Revisi internal 2026-02-26: refactor wiring transaksi/payment tanpa perubahan signature endpoint.
servers:
  - url: /api
    description: Relative API base path
...
paths:
...
  /users/me/profile:
    get:
      tags: [Profile]
      summary: Ambil profil user (namespace /users)
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
    put:
      tags: [Profile]
      summary: Update profil user (namespace /users)
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
      responses:
        '200':
          description: Profil terupdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '422':
          $ref: '#/components/responses/ErrorValidation'
...
  /users/me/devices:
    get:
      tags: [Devices]
      summary: List perangkat user
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Device list
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDevice'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
  /users/me/devices/bind-current:
    post:
      tags: [Devices]
      summary: Bind perangkat current context
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: query
          name: best_effort
          schema:
            type: boolean
          description: Jika true, kegagalan bind dibalas 200 success=false (bukan 403)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                client_ip:
                  type: string
                client_mac:
                  type: string
      responses:
        '200':
          description: Bind result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceBindResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
  /users/me/devices/{device_id}/label:
    put:
      tags: [Devices]
      summary: Update label perangkat
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: device_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
      responses:
        '200':
          description: Label updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
  /users/me/devices/{device_id}:
    delete:
      tags: [Devices]
      summary: Hapus perangkat user
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: device_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  responses:
...
    ErrorUnauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ErrorValidation:
      description: Validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  details:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationErrorDetail'
...
  schemas:
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code
          example: AUTH_UNAUTHORIZED
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
        request_id:
          type: string
    ValidationErrorDetail:
      type: object
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
...
    UserMeResponse:
      type: object
      required: [id, phone_number, full_name, role, approval_status, is_active]
      properties:
        id:
          type: string
          format: uuid
        phone_number:
          type: string
        full_name:
          type: string
        blok:
          type: string
          nullable: true
        kamar:
          type: string
          nullable: true
        is_tamping:
          type: boolean
          nullable: true
        tamping_type:
          type: string
          nullable: true
        role:
          type: string
          enum: [USER, KOMANDAN, ADMIN, SUPER_ADMIN]
        approval_status:
          type: string
          enum: [PENDING_APPROVAL, APPROVED, REJECTED]
        is_active:
          type: boolean
        is_unlimited_user:
          type: boolean
          nullable: true
        total_quota_purchased_mb:
          type: number
          nullable: true
        total_quota_used_mb:
          type: number
          nullable: true
        quota_expiry_date:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
        last_login_at:
          type: string
          format: date-time
          nullable: true
    UserProfileUpdateRequest:
      type: object
      properties:
        full_name:
          type: string
        blok:
          type: string
          nullable: true
        kamar:
          type: string
          nullable: true
        is_tamping:
          type: boolean
          nullable: true
        tamping_type:
          type: string
          nullable: true
    UserDevice:
      type: object
      required: [id, mac_address, is_authorized]
      properties:
        id:
          type: string
          format: uuid
        mac_address:
          type: string
        ip_address:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        is_authorized:
          type: boolean
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
    DeviceBindResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        device:
          $ref: '#/components/schemas/UserDevice'
...